requirements.txt
beautifulsoup4
libclang
tqdm
wheel
pebble
transformers
datasets
evaluate
requests
unidiff